# server/docs/swagger.yaml
basePath: /
definitions:
  model.Api:
    properties:
      category:
        example: 用户管理
        maxLength: 50
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      deleted_at:
        example: "null"
        type: string
      description:
        example: 获取用户列表接口
        maxLength: 255
        type: string
      id:
        example: 1
        type: integer
      method:
        description: GET, POST, PUT, DELETE
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        example: GET
        type: string
      path:
        example: /api/users
        maxLength: 100
        type: string
      status:
        description: 1:启用 2:禁用
        enum:
        - 1
        - 2
        example: 1
        type: integer
      tenant_id:
        description: 租户ID
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - category
    - method
    - path
    type: object
  model.Menu:
    properties:
      component:
        example: User
        maxLength: 100
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      deleted_at:
        example: "null"
        type: string
      hidden:
        example: false
        type: boolean
      icon:
        example: user
        maxLength: 50
        type: string
      id:
        example: 1
        type: integer
      level:
        example: 1
        type: integer
      name:
        example: 用户管理
        maxLength: 50
        minLength: 2
        type: string
      parent_id:
        example: 0
        type: integer
      path:
        example: /user
        maxLength: 100
        type: string
      sort:
        example: 1
        type: integer
      status:
        description: 1:启用 2:禁用
        enum:
        - 1
        - 2
        example: 1
        type: integer
      tenant_id:
        description: 租户ID
        example: 1
        type: integer
      title:
        example: 用户管理
        maxLength: 50
        type: string
      type:
        description: menu:菜单项 group:菜单组
        enum:
        - menu
        - group
        example: menu
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - name
    - title
    type: object
  model.Role:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      deleted_at:
        example: "null"
        type: string
      description:
        example: 系统管理员角色
        maxLength: 255
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        maxLength: 50
        minLength: 2
        type: string
      status:
        description: 1:启用 2:禁用
        enum:
        - 1
        - 2
        example: 1
        type: integer
      tenant_id:
        description: 租户ID
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin + React的后台管理系统API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go-React-Admin API
  version: "1.0"
paths:
  /api/apis:
    get:
      consumes:
      - application/json
      description: 获取所有API的列表
      produces:
      - application/json
      responses:
        "200":
          description: '{"apis":[]model.Api}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"获取API列表失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取API列表
      tags:
      - API管理
    post:
      consumes:
      - application/json
      description: 创建新API接口
      parameters:
      - description: API创建信息
        in: body
        name: api
        required: true
        schema:
          $ref: '#/definitions/model.Api'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"创建API成功","api":model.Api}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"创建API失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建API
      tags:
      - API管理
  /api/apis/{id}:
    delete:
      consumes:
      - application/json
      description: 根据API ID删除API
      parameters:
      - description: API ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"删除API成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"无效的API ID"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"删除API失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除API
      tags:
      - API管理
    put:
      consumes:
      - application/json
      description: 根据API ID更新API信息
      parameters:
      - description: API ID
        in: path
        name: id
        required: true
        type: integer
      - description: API更新信息
        in: body
        name: api
        required: true
        schema:
          $ref: '#/definitions/model.Api'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"更新API成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"更新API失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新API
      tags:
      - API管理
...
  /api/menus:
    get:
      consumes:
      - application/json
      description: 获取所有菜单的列表
      produces:
      - application/json
      responses:
        "200":
          description: '{"menus":[]model.Menu}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"获取菜单列表失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新菜单
      parameters:
      - description: 菜单创建信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"菜单创建成功","menu":model.Menu}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"创建菜单失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /api/menus/{id}:
    delete:
      consumes:
      - application/json
      description: 根据菜单ID删除菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"菜单删除成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"无效的菜单ID"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"删除菜单失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 根据菜单ID更新菜单信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单更新信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"菜单更新成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"更新菜单失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
  /api/menus/user:
    get:
      consumes:
      - application/json
      description: 获取当前用户的菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.Menu,"message":"获取用户菜单成功"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"获取用户菜单失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户菜单
      tags:
      - 菜单管理
...
  /api/roles:
    get:
      consumes:
      - application/json
      description: 获取所有角色的列表
      produces:
      - application/json
      responses:
        "200":
          description: '{"roles":[]model.Role}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"获取角色列表失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: 角色创建信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"角色创建成功","role":model.Role}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"创建角色失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /api/roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据角色ID删除角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"角色删除成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"无效的角色ID"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"删除角色失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 根据角色ID更新角色信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色更新信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"角色更新成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"更新角色失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - 角色管理
...
schemes:
- http
- https
swagger: "2.0"